requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.41
langchain-openai==0.1.25
faiss-cpu==1.9.0.post1
sentence-transformers==3.0.1
streamlit==1.39.0
torch==2.1.0
transformers==4.35.0
pypdf==4.3.1
python-docx==1.1.2
requests==2.32.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.9.0
tiktoken==0.7.0
huggingface-hub==0.24.0
nltk==3.8.1
